tt9

+++ 0x01-python-if_else_loops_functions/9-print_last_digit.py
@@ -1,4 +1,4 @@
 #!/usr/bin/python3
 def print_last_digit(number):
-    ld = number % 10
+    ld =  (int)number % 10
     print("{}".format(ld), end='')
